Make '\s' its own (ignored) token, use better lookahead regexs
The lookahead regexs I was using to match authors, I settled upon during
a time when I didn't understand that whitespace in the regex was
ignored. I was encountering all these bugs, and I decided that my
lookahead regexs weren't working (but things were not working because of
the verbose flag). So put back the sensible lookahead regexs, so that I
don't need to strip the title tokens of their extra parens.

Also, use `\ ` instead of `\s` because `\s` matches newline which is not
what I want.
didn't understand that whitespace inside the regex was ignored